# api/openapi.yaml
openapi: 3.0.3
info:
  title: Rwanda Mineral Data Interoperability Standard API
  description: |
    API specification for the Rwanda Mineral Data Interoperability Standard Version 2.3.
    This API enables interoperability of all systems at national level that handle minerals related data.
    
    Based on the semantic model defined in the Rwanda Mineral Data Interoperability Standard Version 2.3.
    
    **Important**: This standard is designed for data exchange, not as a relational database schema.
  version: 2.3.0
  contact:
    name: Rwanda Mines, Petroleum and Gas Board
    url: https://rmb.rw
  license:
    name: Rwanda Mineral Data Interoperability Standard License
    url: https://rmb.rw/license
servers:
  - url: http://localhost:3000/
    description: Local development server
  - url: https://api-staging.rmb.rw/v1
    description: Rwanda Mineral Data API Staging
  - url: https://api.rmb.rw/v1
    description: Rwanda Mineral Data API Production
...
paths:
  /mine-sites:
    get:
      tags:
        - Mine Sites
      summary: List mine sites
      description: Retrieve a list of mine sites with optional filtering
      operationId: listMineSites
      parameters:
        - name: addressCountry
          in: query
          description: Filter by ICGLR member state code
          schema:
            $ref: '#/components/schemas/ICGLRMemberState'
        - name: certificationStatus
          in: query
          description: Filter by certification status (0=Blue, 1=Green, 2=Yellow, 3=Red)
          schema:
            type: integer
            enum: [0, 1, 2, 3]
        - name: activityStatus
          in: query
          description: Filter by activity status (0=Abandoned, 1=Active, 2=Non-active)
          schema:
            type: integer
            enum: [0, 1, 2]
        - name: mineral
          in: query
          description: Filter by mineral code (HS Code - primary, or IMA Code)
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MineSiteListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Mine Sites
      summary: Create mine site
      description: Create a new mine site record
      operationId: createMineSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MineSite'
          application/ld+json:
            schema:
              $ref: '#/components/schemas/MineSite'
      responses:
        '201':
          description: Mine site created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MineSite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /mine-sites/{icglrId}:
    get:
      tags:
        - Mine Sites
      summary: Get mine site by ICGLR ID
      description: Retrieve a specific mine site by its ICGLR identification number
      operationId: getMineSite
      parameters:
        - name: icglrId
          in: path
          required: true
          description: "ICGLR Mine Site Identification Number (format: CC-[Lat]-[Long]-NNNNN)"
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MineSite'
            application/ld+json:
              schema:
                $ref: '#/components/schemas/MineSite'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Mine Sites
      summary: Update mine site
      description: Update an existing mine site
      operationId: updateMineSite
      parameters:
        - name: icglrId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MineSite'
      responses:
        '200':
          description: Mine site updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MineSite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /export-certificates:
    get:
      tags:
        - Export Certificates
      summary: List export certificates
      description: Retrieve export certificates with filtering options
      operationId: listExportCertificates
      parameters:
        - name: issuingCountry
          in: query
          description: Filter by issuing country
          schema:
            $ref: '#/components/schemas/ICGLRMemberState'
        - name: identifier
          in: query
          description: Filter by certificate serial number
          schema:
            type: string
        - name: lotNumber
          in: query
          description: Filter by lot number
          schema:
            type: string
        - name: typeOfOre
          in: query
          description: Filter by mineral code (HS Code)
          schema:
            type: string
        - name: dateOfIssuanceFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateOfIssuanceTo
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportCertificateListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Export Certificates
      summary: Create export certificate
      description: Submit a new ICGLR Export Certificate
      operationId: createExportCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCertificate'
      responses:
        '201':
          description: Export certificate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportCertificate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /export-certificates/{identifier}:
    get:
      tags:
        - Export Certificates
      summary: Get export certificate by identifier
      description: Retrieve a specific export certificate by its serial number
      operationId: getExportCertificate
      parameters:
        - name: identifier
          in: path
          required: true
          description: Certificate serial number
          schema:
            type: string
        - name: issuingCountry
          in: query
          required: true
          description: Issuing country code (required for unique identification)
          schema:
            $ref: '#/components/schemas/ICGLRMemberState'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportCertificate'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /lots:
    get:
      tags:
        - Chain of Custody
      summary: List lots
      description: Retrieve Chain of Custody lot records with filtering
      operationId: listLots
      parameters:
        - name: mineSiteId
          in: query
          description: Filter by mine site ICGLR ID
          schema:
            type: string
        - name: mineral
          in: query
          description: Filter by mineral code (HS Code)
          schema:
            type: string
        - name: creatorRole
          in: query
          description: Filter by CoC role code (1-8)
          schema:
            type: integer
            enum: [1, 2, 3, 4, 5, 6, 7, 8]
        - name: originatingOperation
          in: query
          description: Filter by originating operation code (1-8)
          schema:
            type: integer
            enum: [1, 2, 3, 4, 5, 6, 7, 8]
        - name: lotNumber
          in: query
          description: Filter by lot number
          schema:
            type: string
        - name: dateRegistrationFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateRegistrationTo
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LotListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Chain of Custody
      summary: Create lot
      description: Submit a new Chain of Custody lot record
      operationId: createLot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lot'
      responses:
        '201':
          description: Lot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /lots/{lotNumber}:
    get:
      tags:
        - Chain of Custody
      summary: Get lot by lot number
      description: Retrieve a specific lot by its lot number
      operationId: getLot
      parameters:
        - name: lotNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lot'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ICGLRMemberState:
      type: string
      enum: [RW]
      description: ICGLR member state ISO 3166-1 alpha-2 code. National rule It will always be RW
...
    MineSiteLocation:
      type: object
      required: [geolocalization]
      properties:
        geolocalization:
          $ref: '#/components/schemas/Geolocalization'
          description: MD.06 Geolocalization - Geographic localization of the mine site in WGS 84 format. In decimal degree format, with 4 decimals
        nationalCadasterLocalization:
          type: string
          description: National cadaster localization - Given in the terms used by the Member State national mining cadastre
        localGeographicDesignation:
          $ref: '#/components/schemas/Address'
          description: Local geographic designation - Given in local geographic terms (province/state, municipality/chefferie/district)
        polygon:
          type: object
          description: 'A polygonal representation of the site, formatted as GeoJSON (RFC 7946) Polygon with coordinates in WGS84 using up to 4 decimals, each position given as [longitude, latitude]. Example: {"type":"Polygon","coordinates":[[[30.0619,-1.9641],[30.0623,-1.9641],[30.0623,-1.9637],[30.0619,-1.9637],[30.0619,-1.9641]]]}'
          required: [type, coordinates]
          properties:
            type:
              type: string
              enum: [Polygon]
              description: GeoJSON type - must be 'Polygon'
            coordinates:
              type: array
              description: GeoJSON Polygon coordinates - array of linear rings, where each ring is an array of [longitude, latitude] positions
              minItems: 1
              items:
                type: array
                description: Linear ring - array of [longitude, latitude] positions. Must have at least 4 positions and should be closed (first and last position identical).
                minItems: 4
                items:
                  type: array
                  description: Position as [longitude, latitude] in WGS84 (first element is longitude, second is latitude)
                  minItems: 2
                  maxItems: 2
                  items:
                    type: number
                    description: Coordinate value. GeoJSON positions are [longitude, latitude]; longitude is -180..180, latitude is -90..90. Precision up to 4 decimals.
                    minimum: -180
                    maximum: 180
                    multipleOf: 0.0001
          additionalProperties: false
        altitude:
          type: number
          description: Average altitude of the mine site. Given in meters, as a number
...
    MineSite:
      type: object
      required: [icglrId, addressCountry, certificationStatus, activityStatus, mineSiteLocation, mineral, license, owner, allowedTags]
      properties:
        icglrId:
          type: string
          pattern: '^[A-Z]{2}-[+-]?[0-9]+\.[0-9]{4}[+-][0-9]+\.[0-9]{4}-[0-9]+$'
          description: "ICGLR Identification number - A unique mine site identification number to be used at the level of ICGLR member states. Will follow the rules established by ICGLR for reporting and should be present if the data record needs to be reported as-is to ICGLR. Format: CC-[Lat]-[Long]-NNNNN"
        addressCountry:
          $ref: '#/components/schemas/ICGLRMemberState'
          description: The country where the Mine site is located. National rule It will always be RW
        nationalId:
          type: string
          description: National Identification Number - A unique mine site identification number at the level of the country
        certificationStatus:
          type: integer
          enum: [0, 1, 2, 3]
          description: 0=Blue (Uninspected), 1=Green (Certified), 2=Yellow (Yellow-Flagged), 3=Red (Un-Certified)
        activityStatus:
          type: integer
          enum: [0, 1, 2]
          description: 0=Abandoned, 1=Active, 2=Non-active
        mineSiteLocation:
          $ref: '#/components/schemas/MineSiteLocation'
          description: MD.08 Mine Site Location
        mineral:
          type: array
          items:
            type: string
          minItems: 1
          description: "The type or types of Mineral(s) produced at the mine. Takes values strictly from the HS Code classification. For designated minerals: 7108.12.00 (Gold), 2609.00.00 (Cassiterite), 2611.00.00 (Wolframite), 2615.90.00 (Columbite-Tantalite/Coltan)"
        license:
          type: array
          items:
            $ref: '#/components/schemas/License'
          minItems: 1
          description: MD.02 License - Group of terms describing the license and its details. Every mine site has a License field, even when the mine site is unlicensed.
        owner:
          $ref: '#/components/schemas/BusinessEntity'
          description: MD.04 Business Entity - Owner
        operator:
          type: array
          items:
            $ref: '#/components/schemas/BusinessEntity'
          description: MD.04 Business Entity - The identification and full details of the Mine Operator/s if different from the owner
        inspection:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
          description: MD.07 Inspection
        statusChange:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistory'
          description: MD.10 Status History - The history of status changes
        allowedTags:
          type: integer
          description: Allowed Tags - Allowed Number of Tags. Approved Tags (list of tag numbers). Specific for Rwanda, which allocates a list of tag numbers to each mine site
        minesLocations:
          type: array
          items:
            $ref: '#/components/schemas/Geolocalization'
          description: Mine Locations - Geographic localization of the mine site, given in the format lat -1.9641, long 30.0619. Given in WGS 84 format. In decimal degree format, with 4 decimals. Specific for Rwanda
    MineSiteListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MineSite'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    ExportCertificate:
      type: object
      required: [issuingCountry, identifier, exporter, importer, lotNumber, typeOfOre, lotWeight, lotWeightUOM, lotGrade, mineralOrigin, customsValue, dateOfShipment, memberStateIssuingAuthority, nameOfVerifier, positionOfVerifier, dateOfVerification, nameOfValidator, dateOfIssuance, dateOfExpiration, certificateFile]
      properties:
        issuingCountry:
          $ref: '#/components/schemas/ICGLRMemberState'
          description: The ISO 3166-1 alpha-2 code of the issuing country. National rule It will always be RW
        identifier:
          type: string
          description: Serial number - A unique serial number identifying the Certificate. A unique combination of digits and letters from the standard Latin alphabet, unique in the country, starting with the ISO 3166-1 alpha-2 – two-letter country code (e.g., RW2324A2)
        exporter:
          $ref: '#/components/schemas/BusinessEntity'
          description: MD.04 Business Entity - Exporter
        importer:
          $ref: '#/components/schemas/BusinessEntity'
          description: MD.04 Business Entity - Importer
        lotNumber:
          type: string
          description: The exporter's unique lot number or export order number for the lot
        designatedMineralDescription:
          type: string
          description: A description of the Designated Mineral, including the type of ore or concentrate, weight and grade of the lot. Free textual description (0..1, optional)
        typeOfOre:
          type: string
          description: Type of ore - MDC.03 Mineral
        lotWeight:
          type: number
          minimum: 0
          exclusiveMinimum: true
          multipleOf: 0.001
          description: Lot Weight - Numeric format, accepting fractional value. Should be a positive number (>0) with a maximum of 3 decimals (0.000)
        lotWeightUOM:
          type: string
          enum: [TNE, KGM]
          description: Lot Weight Unit of Measure - MDC.07 Unit of measurement (TNE=tonne, KGM=kilogram)
        lotGrade:
          type: string
          description: Lot Grade - Grade expression might vary depending on the mineral
        mineralOrigin:
          type: string
          description: The national origin of the material. One or more country identifiers, separated by one space character (e.g., CD BI RW). In case the origin is not known, the character zero (0) will be used. Validation Can be a single 2-letter country code, array separated by semicolon (e.g., BI;TZ;CD), or zero (0) for unknown origin
        customsValue:
          type: string
          description: The declared Customs value in USD of the Lot. Use Currency code of 3 letters (ISO 4217) and value Decimal format, separated by one space character
        dateOfShipment:
          type: string
          format: date
          description: The date the lot is planned to be shipped
        shipmentRoute:
          type: string
          description: The shipment route - ISO 3166 alpha-2 (array of countries separated by a comma without spaces, e.g., CD,TZ,BI) or array of ISO 3166 alpha-2 and ISO 3166 alpha-3 that includes also standardized denominations of regions
        transportCompany:
          type: string
          description: The transport company responsible for transporting the shipment, if known
        memberStateIssuingAuthority:
          type: string
          description: Member State Issuing authority
        nameOfVerifier:
          type: string
          description: Name of the Rwandan representative responsible for verifying the documentation associated the export with the export and recommending the issuance of an ICGLR Regional Certificate
        positionOfVerifier:
          type: string
          description: Position of the Rwandan representative responsible for verifying the documentation associated the export with the export and recommending the issuance of an ICGLR Regional Certificate
        idOfVerifier:
          type: string
          description: Identification number (where available) of the Rwandan representative responsible for verifying the documentation associated the export with the export and recommending the issuance of an ICGLR Regional Certificate
        dateOfVerification:
          type: string
          format: date
          description: The date the lot is verified by the Rwandan representative
        nameOfValidator:
          type: string
          description: The name of the Rwandan representative empowered to countersign the Certificate to render it valid
        dateOfIssuance:
          type: string
          format: date
          description: The date that the Certificate is countersigned (Certificate is valid from this date)
        dateOfExpiration:
          type: string
          format: date
          description: The date that the Certificate expires. Between date of issuance and date of expiration should not be more than 90 days
        certificateFile:
          type: string
          description: "The certificate as file. MIME types: application/pdf, image/png, image/jpeg"
    ExportCertificateListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ExportCertificate'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Lot:
      type: object
      required: [lotNumber, dateRegistration, timeRegistration, creator, mineral, concentration, mass, unitOfMeasurement, creatorRole, price, originatingOperation]
      properties:
        lotNumber:
          type: string
          description: Lot number given by the CoC actor from which the lot originates
        dateRegistration:
          type: string
          format: date
          description: Date of registration - Date when the registration was created
        timeRegistration:
          type: string
          pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9](:[0-5][0-9])?(Z|[+-][0-1][0-9]:[0-5][0-9])?$'
          description: Time of registration - Time (hh:mm:ss) when the registration was created. Should be automatically generated by the software system
        creator:
          $ref: '#/components/schemas/BusinessEntity'
          description: MD.04 Business Entity - Lot Creator - The business entity that created the lot. Equivalent to current custodian.
        mineral:
          type: string
          description: MDC.03 Mineral - Identifier of the mineral contained in the bag
        concentration:
          type: number
          description: Percentual concentration - Approximate mineral concentration (grade), in percent
        mass:
          type: number
          description: Lot weight - Weight of the Lot
        packageType:
          type: string
          description: Type of packaging - Describes how the Lot is packaged (as bag, barrel, container, etc)
        nrOfPackages:
          type: number
          description: Number of packages - In case a lot is composed of more packages
        unitOfMeasurement:
          type: string
          enum: [TNE, KGM]
          description: MDC.07 Unit of measurement - The unit of measure shall be chosen from UN/ECE Recommendation N°. 20
        mineSiteId:
          type: string
          pattern: '^[A-Z]{2}-[+-]?[0-9]+\.[0-9]{4}[+-][0-9]+\.[0-9]{4}-[0-9]+$'
          description: MD.01 Mine Site - Identifier of the mine site where the ore was extracted. REQUIRED when the Originating operation is Production
        miner:
          type: string
          description: Miner's name - The name of the miner who represented the team. REQUIRED when the Originating operation is Production. Specific for Rwanda
        creatorRole:
          type: array
          items:
            type: integer
            enum: [1, 2, 3, 4, 5, 6, 7, 8]
          minItems: 1
          description: MDC.05 CoC Roles - The role(s) of the entity in the CoC
        recipient:
          $ref: '#/components/schemas/BusinessEntity'
          description: MD.04 Business Entity - Lot Recipient - The identification of the customer or recipient
        price:
          type: number
          description: Price per UOM - Required in Rwanda as $/kg
        originatingOperation:
          type: array
          items:
            type: integer
            enum: [1, 2, 3, 4, 5, 6, 7, 8]
          minItems: 1
          description: MDC.06 Originating operations - The operation(s) that determined the creation of the Lot. Can be combined (e.g., 2;3;4)
        inputLot:
          type: array
          items:
            type: object
            properties:
              lotNumber:
                type: string
          description: Lots that form this lot - Collection of lots that were used to aggregate this lot
        tag:
          $ref: '#/components/schemas/Tag'
          description: MD.11 Tag - Associated tag. REQUIRED when the Originating operation is Production
        taxPaid:
          type: array
          items:
            $ref: '#/components/schemas/Tax'
          description: MD.13 Tax - Values and details of all taxes, fees or royalties paid
        dateSealed:
          type: string
          format: date
          description: Date when sealed - The date when the lot is sealed
        dateShipped:
          type: string
          format: date
          description: Date when shipped - The date when the lot is shipped
        purchaseNumber:
          type: string
          description: Purchase number - A unique purchase order number for the external purchase, in case it is applicable
        purchaseDate:
          type: string
          format: date
          description: Purchase date - The date of the purchase, if applicable
        responsibleStaff:
          type: string
          description: Responsible staff - Name of the responsible staff person(s) who accepted the Designated Mineral Lot and verified the origin
        dateIn:
          type: string
          format: date
          description: Date received - The date when the processor took possession of the lot
        transportationMethod:
          type: string
          description: Transportation method - The method by which the Lot was transported from mine site to processor
        transportationRoute:
          type: string
          description: Transportation route - The transportation route used for the shipment of the incoming Lot
        transportCompany:
          type: string
          description: Transport Company - transport company responsible for transporting the shipment
        exportCertificateId:
          type: string
          description: MD.03 Export Certificate - An ICGLR Certificate number, if for export
    LotListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lot'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "VALIDATION_ERROR"
            message: "Request validation failed"
            details:
              field: "icglrId"
              reason: "Invalid format"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "CONFLICT"
            message: "Resource already exists"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "INTERNAL_ERROR"
            message: "An internal error occurred"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - BearerAuth: []
  - ApiKeyAuth: []